requirements.txt
flask
line-bot-sdk
python-dotenv
openai>=1.40
google-generativeai>=0.5.0
gunicorn
request